Backend pagination html and js injections fixed maxSize of file added excess bootstrap files removed

+++ src/main/java/com/dataart/javaschool/newsportal/repository/ArticleRepository.java
@@ -3,5 +3,5 @@ package com.dataart.javaschool.newsportal.repository;
 import com.dataart.javaschool.newsportal.entity.Article;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ArticleRepository extends PagingAndSortingRepository<Article, Long> {
+public interface ArticleRepository extends PagingAndSortingRepository<Article, Integer> {
 }
